requirements.txt
beautifulsoup4==4.12.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
lxml==5.3.0
pandas==2.2.3
phonenumbers==8.13.47
python-dotenv==1.0.1
requests==2.32.3
tldextract==5.1.2
